requirements.txt
anthropic>=0.92.0
openpyxl>=3.1.5
Pillow>=10.4.0
requests>=2.32.0
python-dotenv>=1.0.1
